Try to make osx build happier.
And yes, why ptrdiff_t even exists and why it can be different,
semantically or otherwise, from ssize_t is beyond me...

+++ lib/libvcl/vcc_backend.c
@@ -708,8 +708,8 @@ vcc_ParseDirector(struct vcc *tl)
 	ERRCHK(tl);
 	if (tl->t->e - tl->t->b > 64) {
 		VSB_printf(tl->sb,
-		    "Name of %.*s too long (max 64, is %zd):\n",
-		    PF(t_first), (tl->t->e - tl->t->b));
+		    "Name of %.*s too long (max 64, is %zu):\n",
+		    PF(t_first), (size_t)(tl->t->e - tl->t->b));
 		vcc_ErrWhere(tl, tl->t);
 		return;
 	}
